퀘스트 관련 테이블 데이터 적용 및 로직 변경
1. 수정 중인 내용이기 때문에 주석 처리 된 부분이 남아있음 (코드 정리 필요함)
2. 아이템 테이블 칼럼 이름 변경
3. json 데이터 삭제

+++ Unity_Portfolio/Assets/02.Scripts/Table/SkillTable.cs
@@ -23,8 +23,8 @@ public class SkillTable : ScriptableObject
 		public string Name;
 		public string ResourceName;
 		public float CoolTime;
-		public float MinDamage;
-		public float MaxDamage;
+		public int MinDamage;
+		public int MaxDamage;
 	}
 
 	public void AddData(TableData data)
